% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ApplicationAdjMat}
\alias{ApplicationAdjMat}
\title{Connectivity of European countries based on capital-city distance less than or equal to 820 km (Cshapes R package).}
\format{
A matrix with 28 rows and 28 columns
\describe{
\item{rows}{The 28 European countries in alphabetical order.}
\item{columns}{The 28 European countries in alphabetical order.}
}
}
\usage{
data(ApplicationAdjMat)
}
\description{
Connectivity of European countries based on capital-city distance less than or equal to 820 km (Cshapes R package).
}
