% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ApplicationCounts}
\alias{ApplicationCounts}
\title{Monthly incidence data on invasive meningococcal disease from 28 European countries (European Centre for Disease Prevention and Control (ECDC), 2024)}
\format{
A matrix with 28 rows and 84 columns
\describe{
\item{rows}{The 28 European countries in alphabetical order.}
\item{columns}{The period of Jan 2013 to Dec 2019.}
}
}
\usage{
data(ApplicationCounts)
}
\description{
Monthly incidence data on invasive meningococcal disease from 28 European countries (European Centre for Disease Prevention and Control (ECDC), 2024)
}
