% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{inf.plot}
\alias{inf.plot}
\title{A function to plot inferred trend and seasonal components from Stan/MCMC output.}
\usage{
inf.plot(inf.object)
}
\arguments{
\item{inf.object}{Resulting object from either "infer" or "infer.old" functions.}
}
\value{
Figures
}
\description{
A function to plot inferred trend and seasonal components from Stan/MCMC output.
}
\examples{
set.seed(4);
sim_adjmat<- matrix(0, nrow = 9, ncol = 9)
uppertriang<- c(1,0,1,0,0,0,0,0,1,0,1,0,0,0,0,0,0,1,0,0,0,1,0,1,0,0,1,0,1,0,0,0,1,1,0,1)
gdata::upperTriangle(sim_adjmat, byrow=TRUE)<- uppertriang
gdata::lowerTriangle(sim_adjmat, byrow=FALSE)<- uppertriang
TRUTHS<- simulate(Model = 0, time = 48, adj.matrix = sim_adjmat)
fit<-infer(TRUTHS[[1]],TRUTHS[[2]],Model=0,adjmat=sim_adjmat,Stan=TRUE,ModEvid=TRUE,OutbreakProb=TRUE)
fit<- fit[[1]]  #fit is a list containing the outbreak probabilities as its second element.
inf.plot(inf.object = fit)

}
