% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{mcmc.plot}
\alias{mcmc.plot}
\title{A function for plotting mcmc diagnostic figures.}
\usage{
mcmc.plot(inf.object)
}
\arguments{
\item{inf.object}{Resulting fit object from either "infer", or "infer.old.}
}
\value{
Figures
}
\description{
A function for plotting mcmc diagnostic figures.
}
\examples{
set.seed(4);
sim_adjmat<- matrix(0, nrow = 9, ncol = 9)
uppertriang<- c(1,0,1,0,0,0,0,0,1,0,1,0,0,0,0,0,0,1,0,0,0,1,0,1,0,0,1,0,1,0,0,0,1,1,0,1)
gdata::upperTriangle(sim_adjmat, byrow=TRUE)<- uppertriang
gdata::lowerTriangle(sim_adjmat, byrow=FALSE)<- uppertriang
TRUTHS<- simulate(Model = 0, time = 48, adj.matrix = sim_adjmat)
fit<-infer(TRUTHS[[1]],TRUTHS[[2]],Model=0,adjmat=sim_adjmat,Stan=TRUE,ModEvid=TRUE,OutbreakProb=TRUE)
fit<- fit[[1]]  #if fit is a list containing the outbreak probabilities as its second element.
mcmc.plot(inf.object = fit, OutbreakProb = TRUE)

set.seed(4);
TRUTHS1<- simulate.old(Model = 0, time = 48, adj.matrix = sim_adjmat)
fit1<-infer.old(TRUTHS1[[1]],TRUTHS1[[2]],Model=0,adjmat=sim_adjmat,Stan=TRUE,ModEvid=TRUE,OutbreakProb=TRUE)
fit1<- fit1[[1]]  #fit1 is a list containing the outbreak probabilities as its second element.
mcmc.plot(inf.object = fit1, OutbreakProb = TRUE)

}
