% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginalLoglikelihood.R
\name{ModelEvidence.old}
\alias{ModelEvidence.old}
\title{A function to compute the marginal log-likelihood for model comparison purposes in the spatio-temporal models presented in (Knorr-Held, et.al., 2003).}
\usage{
ModelEvidence.old(y, e_it, adjmat, Model, inf.object, num_samples = 10000)
}
\arguments{
\item{y}{A space-time data matrix (locations on the row, and time on the columns).}

\item{e_it}{A space-time data matrix showing the number of susceptible individuals.}

\item{adjmat}{The adjacency matrix describing the connectivity of spatial locations.}

\item{Model}{The model specification (ranges from 0 to 6).}

\item{inf.object}{Resulting fit object from "infer".}

\item{num_samples}{Number of samples to be used.}
}
\value{
The marginal log-likelihood
}
\description{
A function to compute the marginal log-likelihood for model comparison purposes in the spatio-temporal models presented in (Knorr-Held, et.al., 2003).
}
\examples{
set.seed(4);
sim_adjmat<- matrix(0, nrow = 9, ncol = 9)
uppertriang<- c(1,0,1,0,0,0,0,0,1,0,1,0,0,0,0,0,0,1,0,0,0,1,0,1,0,0,1,0,1,0,0,0,1,1,0,1)
gdata::upperTriangle(sim_adjmat, byrow=TRUE)<- uppertriang
gdata::lowerTriangle(sim_adjmat, byrow=FALSE)<- uppertriang
TRUTHS1<- simulate.old(Model = 0, time = 48, adj.matrix = sim_adjmat)
fit1<-infer.old(TRUTHS1[[1]],TRUTHS1[[2]],Model=0,adjmat=sim_adjmat,Stan=TRUE,ModEvid=FALSE,OutbreakProb=FALSE)
ModelEvidence.old(y=TRUTHS1[[1]],e_it=TRUTHS1[[2]],Model=0,adjmat=sim_adjmat,inf.object=fit1)

}
