% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutbreakProbability.R
\name{OutbreakProbability}
\alias{OutbreakProbability}
\title{Compute posterior marginal probabilities of outbreak using all thinned MCMC samples.}
\usage{
OutbreakProbability(
  y,
  e_it,
  inf.object,
  adjmat,
  Model,
  burn.in = 2000,
  Cyclic = T
)
}
\arguments{
\item{y}{A space-time data matrix (locations on the row, and time on the columns).}

\item{e_it}{A space-time data matrix showing the number of susceptible individuals.}

\item{inf.object}{Resulting object from either "infer" or "infer.old" functions.}

\item{adjmat}{The adjacency matrix describing the connectivity of spatial locations.}

\item{Model}{The model specification (ranges from 0 to 6).}

\item{burn.in}{Burn-in period for MCMC.}

\item{Cyclic}{A logical argument asking whether a cyclic RW1 prior on s_t was used for inference.}
}
\value{
A matrix of outbreak probabilities.
}
\description{
Compute posterior marginal probabilities of outbreak using all thinned MCMC samples.
}
\examples{
set.seed(4);
sim_adjmat<- matrix(0, nrow = 9, ncol = 9)
uppertriang<- c(1,0,1,0,0,0,0,0,1,0,1,0,0,0,0,0,0,1,0,0,0,1,0,1,0,0,1,0,1,0,0,0,1,1,0,1)
gdata::upperTriangle(sim_adjmat, byrow=TRUE)<- uppertriang
gdata::lowerTriangle(sim_adjmat, byrow=FALSE)<- uppertriang
TRUTHS<- simulate(Model = 1, time = 48, adj.matrix = sim_adjmat, T.prob = G(0.2, 0.4), B = 0.65)
fit<-infer(TRUTHS[[1]],TRUTHS[[2]],Model=1,adjmat=sim_adjmat,Stan=TRUE,ModEvid=TRUE,OutbreakProb=FALSE)
OutP<-OutbreakProbability(y=TRUTHS[[1]],e_it=TRUTHS[[2]],inf.object=fit,adjmat=sim_adjmat,Model=1,Cyclic=TRUE)
image(t(OutP))

}
