% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{sim.plot}
\alias{sim.plot}
\title{A function for plotting simulation figures.}
\usage{
sim.plot(sim.object, names = NULL)
}
\arguments{
\item{sim.object}{Resulting simulation object from either "simulate", or "simulate.old".}
}
\value{
Figures
}
\description{
A function for plotting simulation figures.
}
\examples{
set.seed(4);
sim_adjmat<- matrix(0, nrow = 9, ncol = 9)
uppertriang<- c(1,0,1,0,0,0,0,0,1,0,1,0,0,0,0,0,0,1,0,0,0,1,0,1,0,0,1,0,1,0,0,0,1,1,0,1)
gdata::upperTriangle(sim_adjmat, byrow=TRUE)<- uppertriang
gdata::lowerTriangle(sim_adjmat, byrow=FALSE)<- uppertriang
TRUTHS<- simulate(Model = 0, time = 48, adj.matrix = sim_adjmat)
sim.plot(sim.object = TRUTHS)

set.seed(4);
TRUTHS2<- simulate.old(Model = 0, time = 48, adj.matrix = sim_adjmat)
sim.plot(TRUTHS2)

}
