% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate.old}
\alias{simulate.old}
\title{A function to simulate spatio-temporal datasets from the spatio-temporal models presented in (Knorr-Held, et.al., 2003).}
\usage{
simulate.old(
  Model,
  time,
  adj.matrix,
  e_it = matrix(c(rep(c(rpois(time, 5e+05), rpois(time, 1e+06)), 4), rpois(time, 5e+05)),
    byrow = T, ncol = time),
  B = c(0.55, 0.45),
  T.prob = matrix(c(0.9, 0.1, 0.2, 0.8), nrow = 2, byrow = T),
  r = sim.AR2(time, 0.95, 0.065, 0.57),
  s = sim.Seasonals(time),
  u = sim.Spatials(adj.matrix)
)
}
\arguments{
\item{Model}{The model specification (ranges from 0 to 7).}

\item{time}{Represent the time points}

\item{adj.matrix}{The adjacency matrix describing the connectivity of spatial locations.}

\item{e_it}{A space-time data matrix showing the number of susceptible individuals.}

\item{B}{Autoregressive coefficients.}

\item{T.prob}{Transition probability matrix.}

\item{r}{Trend component.}

\item{s}{Seasonal component.}

\item{u}{Spatial component.}
}
\value{
The simulated dataset.
}
\description{
A function to simulate spatio-temporal datasets from the spatio-temporal models presented in (Knorr-Held, et.al., 2003).
}
\examples{
set.seed(4);
sim_adjmat<- matrix(0, nrow = 9, ncol = 9)
uppertriang<- c(1,0,1,0,0,0,0,0,1,0,1,0,0,0,0,0,0,1,0,0,0,1,0,1,0,0,1,0,1,0,0,0,1,1,0,1)
gdata::upperTriangle(sim_adjmat, byrow=TRUE)<- uppertriang
gdata::lowerTriangle(sim_adjmat, byrow=FALSE)<- uppertriang
TRUTHS1<- simulate.old(Model = 0, time = 48, adj.matrix = sim_adjmat)

}
