% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate}
\alias{simulate}
\title{A function to simulate spatio-temporal datasets from the spatio-temporal models presented in (Adeoye, et.al., 2024).}
\usage{
simulate(
  Model,
  time,
  adj.matrix,
  e_it = matrix(c(rep(c(rpois(time, 5e+05), rpois(time, 1e+06)), 4), rpois(time, 5e+05)),
    byrow = T, ncol = time),
  B = c(1.68, 0.2),
  T.prob = matrix(c(0.9, 0.1, 0.2, 0.8), nrow = 2, byrow = T),
  r = sim.RW2(time, sd = 0.009),
  s = sim.Seasonals2(Amplitude = 1.4),
  u = sim.Spatials(adj.matrix)
)
}
\arguments{
\item{Model}{The model specification (ranges from 0 to 7).}

\item{time}{Represent the time points.}

\item{adj.matrix}{The adjacency matrix describing the connectivity of spatial locations.}

\item{e_it}{A space-time data matrix showing the number of susceptible individuals.}

\item{B}{Autoregressive coefficient(s).}

\item{T.prob}{Transition probability matrix.}

\item{r}{Trend component.}

\item{s}{Seasonal component.}

\item{u}{Spatial component.}
}
\value{
Spatio-temporal dataset
}
\description{
A function to simulate spatio-temporal datasets from the spatio-temporal models presented in (Adeoye, et.al., 2024).
}
\examples{
set.seed(4);
sim_adjmat<- matrix(0, nrow = 9, ncol = 9)
uppertriang<- c(1,0,1,0,0,0,0,0,1,0,1,0,0,0,0,0,0,1,0,0,0,1,0,1,0,0,1,0,1,0,0,0,1,1,0,1)
gdata::upperTriangle(sim_adjmat, byrow=TRUE)<- uppertriang
gdata::lowerTriangle(sim_adjmat, byrow=FALSE)<- uppertriang
TRUTHS<- simulate(Model = 0, time = 48, adj.matrix = sim_adjmat)

}
